Add images for scrolling visual
