+++ app/(auth)/_layout.tsx
@@ -1,4 +1,5 @@
 import { Stack } from 'expo-router';
+import { GestureHandlerRootView } from 'react-native-gesture-handler';
 import { useAuth } from '@/contexts/auth-context';
 
 export default function AuthLayout() {
@@ -7,22 +8,23 @@ export default function AuthLayout() {
   const needsLogin = hasServer && serverHealth?.authRequired !== false;
 
   return (
-    <Stack
-      screenOptions={{
-        headerShown: false,
-        gestureEnabled: false,
-      }}
-    >
-      {/* Get Started — visible when no server configured */}
-      <Stack.Protected guard={!hasServer}>
-        <Stack.Screen name="get-started" />
-        <Stack.Screen name="connect" />
-      </Stack.Protected>
+    <GestureHandlerRootView style={{ flex: 1 }}>
+      <Stack
+        screenOptions={{
+          headerShown: false,
+          gestureEnabled: false,
+        }}
+      >
+        {/* Get Started — visible when no server configured */}
+        <Stack.Protected guard={!hasServer}>
+          <Stack.Screen name="get-started" />
+        </Stack.Protected>
 
-      {/* Login screen — visible when server is set but needs auth */}
-      <Stack.Protected guard={needsLogin}>
-        <Stack.Screen name="login" />
-      </Stack.Protected>
-    </Stack>
+        {/* Login screen — visible when server is set but needs auth */}
+        <Stack.Protected guard={needsLogin}>
+          <Stack.Screen name="login" />
+        </Stack.Protected>
+      </Stack>
+    </GestureHandlerRootView>
   );
 }
