+++ app/(auth)/get-started.tsx
@@ -1,8 +1,8 @@
-import { useEffect } from 'react';
-import { Dimensions, Pressable, StyleSheet, Text, View } from 'react-native';
+import { useEffect, useMemo, useRef } from 'react';
+import { Dimensions, Image, ImageSourcePropType, Pressable, StyleSheet, Text, View } from 'react-native';
 import { LinearGradient } from 'expo-linear-gradient';
-import { router } from 'expo-router';
 import { useSafeAreaInsets } from 'react-native-safe-area-context';
+import BottomSheet from '@gorhom/bottom-sheet';
 import Animated, {
   useSharedValue,
   useAnimatedStyle,
@@ -13,6 +13,7 @@ import Animated, {
 import * as Haptics from 'expo-haptics';
 
 import { AurralLogo } from '@/components/AurralLogo';
+import { ConnectSheet } from './components/ConnectSheet';
 import { useColorScheme } from '@/hooks/use-color-scheme';
 import { Colors, Fonts } from '@/constants/theme';
 
@@ -20,56 +21,116 @@ const { width: SCREEN_WIDTH, height: SCREEN_HEIGHT } = Dimensions.get('window');
 const CARD_SIZE = (SCREEN_WIDTH - 48) / 3;
 const CARD_GAP = 8;
 const CARD_TOTAL = CARD_SIZE + CARD_GAP;
-const NUM_ROWS = Math.ceil((SCREEN_HEIGHT * 0.65) / CARD_TOTAL) + 2;
 
-const PALETTE_DARK = [
-  '#3a2e2e', '#2e3a2e', '#2e2e3a', '#3a352e',
-  '#352e3a', '#2e3a35', '#3a2e35', '#2e353a',
-  '#33302e', '#2e3330', '#302e33', '#332e30',
-  '#3a3333', '#333a33', '#33333a', '#3a3a33',
-  '#2e2e2e', '#353535', '#3a3a3a', '#302e35',
-  '#352e30', '#2e3530', '#303530', '#353030',
+const COVERS: ImageSourcePropType[] = [
+  require('@/assets/images/covers/01.jpg'),
+  require('@/assets/images/covers/02.jpg'),
+  require('@/assets/images/covers/03.jpg'),
+  require('@/assets/images/covers/04.jpg'),
+  require('@/assets/images/covers/05.jpg'),
+  require('@/assets/images/covers/06.jpg'),
+  require('@/assets/images/covers/07.jpg'),
+  require('@/assets/images/covers/08.jpg'),
+  require('@/assets/images/covers/09.jpg'),
+  require('@/assets/images/covers/10.jpg'),
+  require('@/assets/images/covers/11.jpg'),
+  require('@/assets/images/covers/12.jpg'),
+  require('@/assets/images/covers/13.jpg'),
+  require('@/assets/images/covers/14.jpg'),
+  require('@/assets/images/covers/15.jpg'),
+  require('@/assets/images/covers/16.jpg'),
+  require('@/assets/images/covers/17.jpg'),
+  require('@/assets/images/covers/18.jpg'),
+  require('@/assets/images/covers/19.jpg'),
+  require('@/assets/images/covers/20.jpg'),
+  require('@/assets/images/covers/21.jpg'),
+  require('@/assets/images/covers/22.jpg'),
+  require('@/assets/images/covers/23.jpg'),
+  require('@/assets/images/covers/24.jpg'),
+  require('@/assets/images/covers/25.jpg'),
+  require('@/assets/images/covers/26.jpg'),
+  require('@/assets/images/covers/27.jpg'),
+  require('@/assets/images/covers/28.jpg'),
+  require('@/assets/images/covers/29.jpg'),
+  require('@/assets/images/covers/30.jpg'),
+  require('@/assets/images/covers/31.jpg'),
+  require('@/assets/images/covers/32.jpg'),
+  require('@/assets/images/covers/33.jpg'),
+  require('@/assets/images/covers/34.jpg'),
+  require('@/assets/images/covers/35.jpg'),
+  require('@/assets/images/covers/36.jpg'),
+  require('@/assets/images/covers/37.jpg'),
+  require('@/assets/images/covers/38.jpg'),
+  require('@/assets/images/covers/39.jpg'),
+  require('@/assets/images/covers/40.jpg'),
 ];
 
-const PALETTE_LIGHT = [
-  '#d4cfc4', '#c4d4c4', '#c4c4d4', '#d4d0c4',
-  '#d0c4d4', '#c4d4d0', '#d4c4d0', '#c4d0d4',
-  '#cccac4', '#c4ccca', '#cac4cc', '#ccc4ca',
-  '#d4cccc', '#ccd4cc', '#ccccd4', '#d4d4cc',
-  '#c4c4c4', '#cacaca', '#d0d0d0', '#cac4d0',
-  '#d0c4ca', '#c4d0ca', '#cad0ca', '#d0caca',
-];
-
-function generateCards(palette: string[]) {
-  return Array.from({ length: 3 * NUM_ROWS * 2 }, (_, i) => ({
-    id: i,
-    color: palette[i % palette.length],
-  }));
+/**
+ * Returns a new array containing the elements of the input array shuffled in a pseudo-random order,
+ * determined by the provided seed value.
+ *
+ * @param arr The input array to be shuffled.
+ * @param seed A number used as the seed for the pseudo-random number generator, ensuring deterministic shuffling.
+ * @return A new array with the elements shuffled in a pseudo-random order.
+ */
+function shuffled<T>(arr: T[], seed: number): T[] {
+  const a = [...arr];
+  let s = seed;
+  for (let i = a.length - 1; i > 0; i--) {
+    s = (s * 16807) % 2147483647;
+    const j = s % (i + 1);
+    [a[i], a[j]] = [a[j], a[i]];
+  }
+  return a;
 }
 
+/**
+ * Generates a two-dimensional array of image sources divided into columns.
+ * The images are shuffled based on the provided seed value.
+ *
+ * @param {number} seed - The seed value used to shuffle the array of image sources.
+ * @return {ImageSourcePropType[][]} A two-dimensional array where each sub-array represents a column of image sources.
+ */
+function generateColumns(seed: number): ImageSourcePropType[][] {
+  const all = shuffled(COVERS, seed);
+  const perCol = Math.floor(all.length / 3);
+  return [0, 1, 2].map((colIdx) => {
+    return all.slice(colIdx * perCol, (colIdx + 1) * perCol);
+  });
+}
+
+/**
+ * Renders a vertically scrolling column of images with animation.
+ *
+ * @param {Object} props - The properties object.
+ * @param {ImageSourcePropType[]} props.images - An array of image sources to display in the scrolling column.
+ * @param {number} props.speed - The speed of the scrolling animation in milliseconds.
+ * @param {number} props.offset - An additional offset to apply to the initial vertical position.
+ * @return The animated scrolling column component.
+ */
 function ScrollingColumn({
-  cards,
+  images,
   speed,
   offset,
 }: {
-  cards: { id: number; color: string }[];
+  images: ImageSourcePropType[];
   speed: number;
   offset: number;
 }) {
   const translateY = useSharedValue(0);
-  const totalHeight = (NUM_ROWS * CARD_TOTAL);
+  const setHeight = images.length * CARD_TOTAL;
 
   useEffect(() => {
     translateY.value = 0;
     translateY.value = withRepeat(
-      withTiming(-totalHeight, {
+      withTiming(-setHeight, {
         duration: speed,
         easing: Easing.linear,
       }),
       -1,
       false,
     );
-  }, [translateY, totalHeight, speed]);
+  }, [translateY, setHeight, speed]);
 
   const animatedStyle = useAnimatedStyle(() => ({
     transform: [{ translateY: translateY.value + offset }],
@@ -77,42 +138,39 @@ function ScrollingColumn({
 
   return (
     <Animated.View style={[styles.column, animatedStyle]}>
-      {cards.map((card) => (
-        <View
-          key={card.id}
-          style={[
-            styles.card,
-            { backgroundColor: card.color },
-          ]}
-        />
-      ))}
+      {[0, 1].map((set) =>
+        images.map((source, i) => (
+          <Image
+            key={`${set}-${i}`}
+            source={source}
+            style={styles.card}
+          />
+        )),
+      )}
     </Animated.View>
   );
 }
 
 export default function GetStartedScreen() {
-  const colorScheme = useColorScheme() ?? 'light';
+  const colorScheme = useColorScheme();
   const colors = Colors[colorScheme];
   const insets = useSafeAreaInsets();
-  const palette = colorScheme === 'dark' ? PALETTE_DARK : PALETTE_LIGHT;
-  const cards = generateCards(palette);
+  const bottomSheetRef = useRef<BottomSheet>(null);
 
-  const col1 = cards.slice(0, NUM_ROWS * 2);
-  const col2 = cards.slice(NUM_ROWS * 2, NUM_ROWS * 4);
-  const col3 = cards.slice(NUM_ROWS * 4, NUM_ROWS * 6);
+  const [col1, col2, col3] = useMemo(() => generateColumns(42), []);
 
-  const handleGetStarted = () => {
-    Haptics.impactAsync(Haptics.ImpactFeedbackStyle.Light);
-    router.push('/(auth)/connect');
+  const handleGetStarted = async () => {
+    await Haptics.impactAsync(Haptics.ImpactFeedbackStyle.Light);
+    bottomSheetRef.current?.expand();
   };
 
   return (
     <View style={[styles.container, { backgroundColor: colors.background }]}>
-      {/* Scrolling cards background */}
+      {/* Scrolling album art */}
       <View style={styles.cardsContainer}>
-        <ScrollingColumn cards={col1} speed={60000} offset={0} />
-        <ScrollingColumn cards={col2} speed={45000} offset={-CARD_TOTAL / 2} />
-        <ScrollingColumn cards={col3} speed={55000} offset={-CARD_TOTAL / 3} />
+        <ScrollingColumn images={col1} speed={120000} offset={0} />
+        <ScrollingColumn images={col2} speed={90000} offset={-CARD_TOTAL / 2} />
+        <ScrollingColumn images={col3} speed={110000} offset={-CARD_TOTAL / 3} />
       </View>
 
       {/* Gradient overlay */}
@@ -144,6 +202,8 @@ export default function GetStartedScreen() {
           </Text>
         </Pressable>
       </View>
+
+      <ConnectSheet ref={bottomSheetRef} />
     </View>
   );
 }
